requirements.txt
flask>=3.0.0
jinja2>=3.1.2
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-openai>=0.1.16
langchain-community>=0.2.10
langgraph==0.5.4
pydantic>=2.7
duckduckgo-search>=5.3.0
ddgs
lxml
faiss-cpu>=1.8.0
yfinance>=0.2.40
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
numpy>=1.26.0
rapidfuzz>=3.9.0
pytest>=8.1.0
pytest-cov>=5.0.0
requests>=2.31.0
